arreglo tests publicos T1

+++ Tareas/T1/tests_publicos/test_05_guardar_estado.py
@@ -1,16 +1,19 @@
-import unittest, os
+import unittest, os, sys
 from os import path
 
 from dccasillas import DCCasillas
 
 from tests_publicos.timeout_function import timeout
 
-N_SECOND = 5
+sys.stdout = open(os.devnull, 'w')
+N_SECOND = 10
 
 class TestGuardarEstado(unittest.TestCase):
     def shortDescription(self):
         doc = self._testMethodDoc
         return doc or None
+    
+    archivos = []
     
     @timeout(N_SECOND)
     def test_00_guardar_estado_True_1(self):
@@ -21,6 +24,7 @@ class TestGuardarEstado(unittest.TestCase):
         config = "config01.txt"
         user = "Agusrush21"
         archivo = user + ".txt"
+        self.archivos.append(archivo)
         
         ruta_esperada = path.join("data", archivo)
         
@@ -32,14 +36,17 @@ class TestGuardarEstado(unittest.TestCase):
         self.assertTrue(path.exists(ruta_esperada), "No existe el archivo")
         self.assertTrue(resultado_estudiante, "Debe retornar True")
 
+
     @timeout(N_SECOND)
     def test_01_guardar_estado_True_2(self):
         """
         Verifica que se guarde correctamente el archivo, devuelva True.
         """
+
         config = "config01.txt"
         user = "cruz"
         archivo = user + ".txt"
+        self.archivos.append(archivo)
         
         ruta_esperada = path.join("data", archivo)
         
@@ -58,9 +65,11 @@ class TestGuardarEstado(unittest.TestCase):
         Verifica que se guarde correctamente el archivo, devuelva True.
         Ademas evalua el puntaje y movimientos.
         """
+
         config = "config01.txt"
         user = "R3Y3X"
         archivo = user + ".txt"
+        self.archivos.append(archivo)
         
         ruta_esperada = path.join("data", archivo)
         
@@ -81,9 +90,11 @@ class TestGuardarEstado(unittest.TestCase):
         Verifica que se guarde correctamente el archivo, devuelva True.
         Ademas evalua el puntaje y movimientos.
         """
+
         config = "config01.txt"
         user = "trinicamposl"
         archivo = user + ".txt"
+        self.archivos.append(archivo)
         
         ruta_esperada = path.join("data", archivo)
         
@@ -105,9 +116,11 @@ class TestGuardarEstado(unittest.TestCase):
         Ademas evalua el puntaje y movimientos.
         Evalua distintos cambios en las casillas.
         """
+
         config = "config01.txt"
         user = "pepe"
         archivo = user + ".txt"
+        self.archivos.append(archivo)
     
         ruta_esperada = path.join("data", archivo)
         
@@ -121,10 +134,10 @@ class TestGuardarEstado(unittest.TestCase):
         self.assertTrue(path.exists(ruta_esperada), "No existe el archivo")
         self.assertTrue(resultado_estudiante, "Debe retornar True")
 
-        dccasillas_estudiante.tableros[4].tablero[0][0] = "." + dccasillas_estudiante.tableros[4].tablero[0][0]
-        dccasillas_estudiante.tableros[4].tablero[2][0] = "." + dccasillas_estudiante.tableros[4].tablero[0][0]
-        dccasillas_estudiante.tableros[4].tablero[3][1] = "." + dccasillas_estudiante.tableros[4].tablero[0][0]
-        dccasillas_estudiante.tableros[4].tablero[7][0] = "." + dccasillas_estudiante.tableros[4].tablero[0][0]
+        dccasillas_estudiante.tableros[4].tablero[0][0] = "X" + dccasillas_estudiante.tableros[4].tablero[0][0]
+        dccasillas_estudiante.tableros[4].tablero[2][0] = "X" + dccasillas_estudiante.tableros[4].tablero[2][0]
+        dccasillas_estudiante.tableros[4].tablero[3][1] = "X" + dccasillas_estudiante.tableros[4].tablero[3][1]
+        dccasillas_estudiante.tableros[4].tablero[7][0] = "X" + dccasillas_estudiante.tableros[4].tablero[7][0]
 
         resultado_estudiante = dccasillas_estudiante.guardar_estado()
 
@@ -139,9 +152,11 @@ class TestGuardarEstado(unittest.TestCase):
         Ademas evalua el puntaje y movimientos.
         Evalua distintos cambios en las casillas
         """
+
         config = "config01.txt"
         user = "maria"
         archivo = user + ".txt"
+        self.archivos.append(archivo)
     
         ruta_esperada = path.join("data", archivo)
         
@@ -155,16 +170,16 @@ class TestGuardarEstado(unittest.TestCase):
         self.assertTrue(path.exists(ruta_esperada), "No existe el archivo")
         self.assertTrue(resultado_estudiante, "Debe retornar True")
 
-        dccasillas_estudiante.tableros[2].tablero[0][3] = "." + dccasillas_estudiante.tableros[4].tablero[0][0]
-        dccasillas_estudiante.tableros[2].tablero[1][2] = "." + dccasillas_estudiante.tableros[4].tablero[0][0]
-        dccasillas_estudiante.tableros[2].tablero[2][1] = "." + dccasillas_estudiante.tableros[4].tablero[0][0]
-        dccasillas_estudiante.tableros[2].tablero[0][0] = "." + dccasillas_estudiante.tableros[4].tablero[0][0]
+        dccasillas_estudiante.tableros[2].tablero[0][3] = "X" + dccasillas_estudiante.tableros[4].tablero[0][3]
+        dccasillas_estudiante.tableros[2].tablero[1][2] = "X" + dccasillas_estudiante.tableros[4].tablero[1][2]
+        dccasillas_estudiante.tableros[2].tablero[2][1] = "X" + dccasillas_estudiante.tableros[4].tablero[2][1]
+        dccasillas_estudiante.tableros[2].tablero[0][0] = "X" + dccasillas_estudiante.tableros[4].tablero[0][0]
 
         resultado_estudiante = dccasillas_estudiante.guardar_estado()
 
-        dccasillas_estudiante.tableros[2].tablero[0][3] = dccasillas_estudiante.tableros[4].tablero[0][0].replace(".","")
-        dccasillas_estudiante.tableros[2].tablero[1][2] = "." + dccasillas_estudiante.tableros[4].tablero[0][0].replace(".","")
-        dccasillas_estudiante.tableros[2].tablero[2][1] = "." + dccasillas_estudiante.tableros[4].tablero[0][0].replace(".","")
+        dccasillas_estudiante.tableros[2].tablero[0][3] = dccasillas_estudiante.tableros[4].tablero[0][3].replace("X","")
+        dccasillas_estudiante.tableros[2].tablero[1][2] = "X" + dccasillas_estudiante.tableros[4].tablero[1][2].replace("X","")
+        dccasillas_estudiante.tableros[2].tablero[2][1] = "X" + dccasillas_estudiante.tableros[4].tablero[2][1].replace("X","")
 
         resultado_estudiante = dccasillas_estudiante.guardar_estado()
 
@@ -172,11 +187,44 @@ class TestGuardarEstado(unittest.TestCase):
         self.assertTrue(resultado_estudiante, "Debe retornar True")
     
     
+    @timeout(N_SECOND)
+    def test_06_guardar_estado_False_1(self):
+        """
+        Verifica que no se guarde correctamente el archivo. devuelva False
+        Nombre de usuario invalido.
+        """
+
+        config = "config01.txt"
+        user = ""
+
+        dccasillas_estudiante = DCCasillas(usuario=user, config=config)
+        resultado_estudiante = dccasillas_estudiante.guardar_estado()
+
+        self.assertFalse(resultado_estudiante, "Debe retornar False")
+
+    
+    @timeout(N_SECOND)
+    def test_07_guardar_estado_False_2(self):
+        """
+        Verifica que no se guarde correctamente el archivo. devuelva False
+        No hay tableros que guardar.
+        """
+
+        config = "config02.txt"
+        user = "Frandeus"
+
+        dccasillas_estudiante = DCCasillas(usuario=user, config=config)
+        dccasillas_estudiante.tableros = []
+        resultado_estudiante = dccasillas_estudiante.guardar_estado()
+
+        self.assertFalse(resultado_estudiante, "Debe retornar False")
+
+
     @classmethod
     def tearDownClass(cls):
         directorio = "data"
         for archivo in os.listdir(directorio):
             if archivo.endswith(".txt"):
                 ruta_completa = path.join(directorio, archivo)
-                if path.isfile(ruta_completa):
+                if path.isfile(ruta_completa) and archivo in cls.archivos:
                     os.remove(ruta_completa)
